+++ app/models/carto/user.rb
@@ -46,8 +46,8 @@ class Carto::User < ActiveRecord::Base
 
   delegate [
       :database_username, :database_password, :in_database,
-      :db_size_in_bytes, :get_api_calls, :table_count, :public_visualization_count, :visualization_count,
-      :twitter_imports_count, :maps_count
+      :db_size_in_bytes, :get_api_calls, :table_count, :public_visualization_count, :all_visualization_count,
+      :visualization_count, :twitter_imports_count
     ] => :service
 
   attr_reader :password
@@ -141,7 +141,7 @@ class Carto::User < ActiveRecord::Base
   end
 
   def remove_logo?
-    /(FREE|MAGELLAN|JOHN SNOW|ACADEMY|ACADEMIC|ON HOLD)/i.match(self.account_type) ? false : true
+    Carto::AccountType.new.remove_logo?(self)
   end
 
   def organization_username
@@ -293,12 +293,7 @@ class Carto::User < ActiveRecord::Base
   end
 
   def soft_geocoding_limit?
-    if self[:soft_geocoding_limit].nil?
-      plan_list = "ACADEMIC|Academy|Academic|INTERNAL|FREE|AMBASSADOR|ACADEMIC MAGELLAN|PARTNER|FREE|Magellan|Academy|ACADEMIC|AMBASSADOR"
-      (self.account_type =~ /(#{plan_list})/ ? false : true)
-    else
-      self[:soft_geocoding_limit]
-    end
+    Carto::AccountType.new.soft_geocoding_limit?(self)
   end
   alias_method :soft_geocoding_limit, :soft_geocoding_limit?
 
@@ -325,11 +320,7 @@ class Carto::User < ActiveRecord::Base
   end
 
   def dedicated_support?
-    /(FREE|MAGELLAN|JOHN SNOW|ACADEMY|ACADEMIC|ON HOLD)/i.match(self.account_type) ? false : true
-  end
-
-  def remove_logo?
-    /(FREE|MAGELLAN|JOHN SNOW|ACADEMY|ACADEMIC|ON HOLD)/i.match(self.account_type) ? false : true
+    Carto::AccountType.new.dedicated_support?(self)
   end
 
   def arcgis_datasource_enabled?