+++ app/models/data_import.rb
@@ -66,7 +66,8 @@ class DataImport < Sequel::Model
     'original_url',
     'privacy',
     'http_response_code',
-    'rejected_layers'
+    'rejected_layers',
+    'runner_warnings'
   ]
 
   # This attributes will get removed from public_values upon calling api_call_public_values
@@ -431,7 +432,7 @@ class DataImport < Sequel::Model
     new_table_name = import_from_query(table_name, query)
     sanitize_columns(new_table_name)
 
-    self.update(table_names: new_table_name)
+    self.update(table_names: new_table_name, service_name: nil)
     migrate_existing(new_table_name)
 
     self.results.push CartoDB::Importer2::Result.new(success: true, error: nil)
@@ -649,6 +650,7 @@ class DataImport < Sequel::Model
       self.results    = importer.results
       self.error_code = importer.error_code
       self.rejected_layers = importer.rejected_layers.join(',') if !importer.rejected_layers.empty?
+      self.runner_warnings = runner.warnings.to_json if !runner.warnings.empty?
 
       # http_response_code is only relevant if a direct download is performed
       if !runner.nil? && datasource_provider.providers_download_url?
@@ -809,8 +811,9 @@ class DataImport < Sequel::Model
   end
 
   def decorate_log(data_import)
-    decoration = { retrieved_items: 0}
-    if data_import.success && data_import.table_id
+    decoration = { retrieved_items: 0 }
+    if data_import.success && data_import.table_id && data_import.migrate_table.nil? && data_import.from_query.nil? &&
+       data_import.table_copy.nil?
       datasource = get_datasource_provider
       if datasource.persists_state_via_data_import?
         decoration = datasource.get_audit_stats