+++ lib/static_maps_url_helper.rb
@@ -4,38 +4,33 @@ module Carto
 
     def url_for_static_map(request, visualization, map_width, map_height)
       username = CartoDB.extract_subdomain(request)
-      request_protocol = request.protocol.sub('://','')
-      static_maps_base_url(username, request_protocol) + static_maps_image_url_fragment(visualization.id, map_width, map_height)
+      request_protocol = request.protocol.sub('://', '')
+      static_maps_base_url(username, request_protocol) +
+        static_maps_image_url_fragment(visualization.id, map_width, map_height)
     end
 
     def url_for_static_map_without_request(username, request_protocol, visualization, map_width, map_height)
-      static_maps_base_url(username, request_protocol) + static_maps_image_url_fragment(visualization.id, map_width, map_height)
+      static_maps_base_url(username, request_protocol) +
+        static_maps_image_url_fragment(visualization.id, map_width, map_height)
     end
 
     private
 
     # INFO: Assumes no trailing '/' comes inside, so returned string doesn't has it either
     def static_maps_base_url(username, request_protocol)
-      config = get_static_maps_api_cdn_config
-
+      config = get_cdn_config
       if !config.nil? && !config.empty?
-        # Sample formats:
-        # {protocol}://{user}.cartodb.com
-        # {protocol}://zone.cartocdn.com/{user}
-        base_url = config
+        "#{request_protocol}://#{config[request_protocol]}/#{username}"
       else
-        # Typical format (but all parameters except {user} come already replaced): 
-        # {protocol}://{user}.{maps_domain}:{port}/
-        base_url = ApplicationHelper.maps_api_template('public')
+        # sample format: {protocol}://{user}.{maps_domain}:{port}/
+        # All parameters except {user} come already replaced
+        ApplicationHelper.maps_api_template('public').sub('{user}', username)
       end
-
-      base_url.sub('{protocol}', CartoDB.protocol(request_protocol))
-              .sub('{user}', username)
     end
 
     # INFO: To ease testing while we keep the config in a global array...
-    def get_static_maps_api_cdn_config
-      Cartodb.config[:maps_api_cdn_template]
+    def get_cdn_config
+      Cartodb.config[:cdn_url]
     end
 
     def static_maps_image_url_fragment(visualization_id, width, height)
@@ -45,5 +40,4 @@ module Carto
     end
 
   end
-
 end