+++ services/datasources/lib/datasources/url/instagram_oauth.rb
@@ -23,12 +23,7 @@ module CartoDB
         # @throws UninitializedError
         # @throws MissingConfigurationError
         def initialize(config, user)
-          super
-
-          raise UninitializedError.new('missing user instance', DATASOURCE_NAME)        if user.nil?
-          raise MissingConfigurationError.new('missing app_key', DATASOURCE_NAME)       unless config.include?('app_key')
-          raise MissingConfigurationError.new('missing app_secret', DATASOURCE_NAME)    unless config.include?('app_secret')
-          raise MissingConfigurationError.new('missing callback_url', DATASOURCE_NAME)  unless config.include?('callback_url')
+          super(config, user, %w{ app_key app_secret callback_url }, DATASOURCE_NAME)
 
           @user         = user
           @app_key      = config.fetch('app_key')