+++ spec/support/helpers.rb
@@ -26,13 +26,15 @@ module HelperMethods
   end
 
   def serve_file(file_path, options = {})
+    port = get_unused_port
+
     require 'webrick'
     server = WEBrick::HTTPServer.new(
-      :AccessLog       => [],
-      :Logger          => WEBrick::Log::new("/dev/null", 7), #comment this line if weird things happen
-      :Port            => 9779,
-      :DocumentRoot    => File.dirname(file_path),
-      :RequestCallback => Proc.new() { |req, res|
+      AccessLog: [],
+      Logger: WEBrick::Log::new("/dev/null", 7), # comment this line if weird things happen
+      Port: port,
+      DocumentRoot: File.dirname(file_path),
+      RequestCallback: Proc.new() { |_req, res|
         options[:headers].each { |k, v| res[k] = v } if options[:headers].present?
         if options[:headers].present? && options[:headers]['content-type'].present?
           res.content_type = options[:headers]['content-type']
@@ -45,7 +47,7 @@ module HelperMethods
     a = Thread.new { server.start }
 
     begin
-      yield "http://localhost:9779/#{File.basename(file_path)}" if block_given?
+      yield "http://localhost:#{port}/#{File.basename(file_path)}" if block_given?
     rescue => e
       raise e
     ensure
@@ -56,6 +58,16 @@ module HelperMethods
     end
   end
 
+  def get_unused_port
+    used_ports_command = `netstat -tln | tail -n +3 | awk '{ print $4 }' | cut -f2 -d ':'`
+    used_ports = used_ports_command.split("\n").map(&:to_i)
+
+    (10000..65535).each do |port|
+      return port if !used_ports.include?(port)
+    end
+
+    raise "No ports available on machine."
+  end
 
   def get_json(path, params = {}, headers ={}, &block)
     get path, params, headers